feat: Add DeviceInfo interface to feature.ts oc: 4273 (#2)
* feat: Add DeviceInfo interface to feature.ts

This commit adds the DeviceInfo interface from the '@capacitor/device' package to the feature.ts file. This interface includes properties for appVersion and os, which will provide information about the device's operating system and application version.

* chore: Add appVersion property to WmDeviceInfo interface

This commit adds the appVersion property to the WmDeviceInfo interface in order to store the version of the application.